refactor: re structure api files and folder
