+++ src/api/cart/addLineItem.ts
@@ -1,5 +1,4 @@
-import Medusa from "@medusajs/medusa-js";
-
+import medusa from "@api/medusa";
 export type Item = {
   cardId: string;
   variant_id: string;
@@ -7,10 +6,6 @@ export type Item = {
 };
 
 export const addLineItem = async ({ cardId, variant_id, quantity }: Item) => {
-  const medusa = new Medusa({
-    baseUrl: import.meta.env.PUBLIC_BASE_URL,
-    maxRetries: 3,
-  });
   return medusa.carts.lineItems.create(cardId, {
     variant_id,
     quantity,
