+++ src/api/medusa/index.ts
@@ -1,0 +1,8 @@
+import Medusa from "@medusajs/medusa-js";
+
+const medusa = new Medusa({
+  baseUrl: import.meta.env.PUBLIC_BASE_URL,
+  maxRetries: 3,
+});
+
+export default medusa;
