+++ src/api/product/getProduct.ts
@@ -1,0 +1,4 @@
+import medusa from "@api/medusa";
+export const getProduct = async (productId: string) => {
+  return medusa.products.retrieve(productId);
+};
