+++ src/api/product/listProducts.ts
@@ -1,0 +1,5 @@
+import medusa from "@api/medusa";
+
+export const listProducts = async () => {
+  return medusa.products.list();
+};
